my_fastapi_project/pyproject.toml
[tool.poetry]
name = "my-fastapi-project"
version = "0.1.0"
description = ""
authors = ["ddinhcchi <ddinhcchi540@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.1"}
sqlalchemy = "^2.0.28"
databases = "^0.9.0"
alembic = "^1.13.1"
pydantic = {extras = ["email"], version = "^2.6.4"}
passlib = "^1.7.4"
python-jose = "^3.3.0"
psycopg2 = "^2.9.9"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
